requirements.txt
asgiref==3.11.0
attrs==25.4.0
coverage==7.13.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==7.0.0
pytest-django==4.7.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.6.0
